+++ tcx-primitive/src/sr25519.rs
@@ -1,19 +1,10 @@
 use crate::ecc::{KeyError, PrivateKey as TraitPrivateKey, PublicKey as TraitPublicKey};
-use crate::Result;
+use crate::{FromHex, Result, ToHex};
+use schnorrkel::{ExpansionMode, MiniSecretKey};
 use sp_core::sr25519::{Pair, Public};
 use sp_core::{Pair as TraitPair, Public as TraitPublic};
 use sp_keyring::ed25519::Keyring;
 use std::io;
-//
-//fn transform_mnemonic_error(err: sp_core::crypto::SecretStringError) -> failure::Error {
-//
-//    match err {
-//        SecretStringError:: => Error::MnemonicChecksumInvalid,
-//        bip39::ErrorKind::InvalidWord => Error::MnemonicWordInvalid,
-//        bip39::ErrorKind::InvalidWordLength(_) => Error::MnemonicLengthInvalid,
-//        _ => Error::MnemonicInvalid,
-//    }
-//}
 
 #[derive(Clone)]
 pub struct Sr25519PublicKey(pub Public);
@@ -33,38 +24,19 @@ impl From<Pair> for Sr25519PrivateKey {
     }
 }
 
-//impl Sr25519PublicKey {
-//    pub fn to_compressed(&self) -> Vec<u8> {
-//        self.0.key.serialize().to_vec()
-//    }
-//
-//    pub fn to_uncompressed(&self) -> Vec<u8> {
-//        self.0.key.serialize_uncompressed().to_vec()
-//    }
-//}
-
-//impl Sr25519PrivateKey {
-//    pub fn from_wif(wif: &str) -> Result<Self> {
-//        Secp256k1PrivateKey::from_ss58check(wif)
-//    }
-//}
-
-impl Sr25519PrivateKey {
-    pub fn from_mnemonic(mnemonic: &str) -> Result<Self> {
-        let pair = Pair::from_phrase(mnemonic, None).map_err(|_| format_err!("mnemonic_error"))?;
-        Ok(Sr25519PrivateKey(pair.0))
-    }
-}
 impl TraitPrivateKey for Sr25519PrivateKey {
     type PublicKey = Sr25519PublicKey;
 
     fn from_slice(data: &[u8]) -> Result<Self> {
-        unimplemented!()
+        let mini_key: MiniSecretKey =
+            MiniSecretKey::from_bytes(data).expect("32 bytes can always build a key; qed");
+
+        let kp = mini_key.expand_to_keypair(ExpansionMode::Ed25519);
+        Ok(Sr25519PrivateKey(Pair::from(kp)))
     }
 
     fn public_key(&self) -> Self::PublicKey {
-        // todo:
-        Sr25519PublicKey::from_slice(self.0.public().as_slice()).unwrap()
+        Sr25519PublicKey(self.0.public())
     }
 
     fn sign(&self, data: &[u8]) -> Result<Vec<u8>> {
@@ -81,7 +53,8 @@ impl TraitPrivateKey for Sr25519PrivateKey {
         //        let (recover_id, sign) = signature.serialize_compact();
         //        let signed_bytes = [sign[..].to_vec(), vec![(recover_id.to_i32()) as u8]].concat();
         //        Ok(signed_bytes)
-        unimplemented!()
+        // https://www.deadalnix.me/2017/02/17/schnorr-signatures-for-not-so-dummies/
+        self.sign(data)
     }
 
     fn to_bytes(&self) -> Vec<u8> {
@@ -100,12 +73,22 @@ impl TraitPublicKey for Sr25519PublicKey {
         Ok(Sr25519PublicKey(Public::from_slice(data)))
     }
 
-    fn write_into<W: io::Write>(&self, writer: W) {
-        unimplemented!()
-    }
-
     fn to_bytes(&self) -> Vec<u8> {
         self.0.to_vec()
+    }
+}
+
+impl ToHex for Sr25519PublicKey {
+    fn to_hex(&self) -> String {
+        hex::encode(self.0.to_raw_vec())
+    }
+}
+
+impl FromHex for Sr25519PublicKey {
+    fn from_hex(hex: &str) -> Result<Self> {
+        let bytes = hex::decode(hex)?;
+        let pk = Sr25519PublicKey::from_slice(bytes.as_slice())?;
+        Ok(pk)
     }
 }
 
@@ -143,19 +126,3 @@ impl TraitPublicKey for Sr25519PublicKey {
 //    }
 //}
 //
-//pub fn verify_private_key(private_key: &str, coin: &CoinInfo) -> Result<String> {
-//    if let Some(network) = network_from_coin(coin) {
-//        let (pk, version) = Secp256k1PrivateKey::from_ss58check_with_version(private_key)?;
-//        if version[0] != network.private_prefix {
-//            return Err(KeyError::InvalidPrivateKey.into());
-//        } else {
-//            return Ok(hex::encode(pk.to_bytes()));
-//        }
-//    }
-//    Ok(private_key.to_string())
-//}
-//
-//pub fn private_key_without_version(private_key: &str) -> Result<Vec<u8>> {
-//    let (pk, _version) = Secp256k1PrivateKey::from_ss58check_with_version(private_key)?;
-//    Ok(pk.to_bytes())
-//}
